Cleanup

+++ ledger-kafka/src/test/java/com/obsidiandynamics/blackstrom/ledger/TestTopic.java
@@ -22,6 +22,6 @@ public final class TestTopic {
   }
   
   public static NewTopic newOf(String topic) {
-    return new NewTopic(topic, (short) 1, (short) 1);
+    return new NewTopic(topic, 1, (short) 1);
   }
 }
